Modal changing, HOC deleted
