+++ src/components/modal/modal.module.css
@@ -8,3 +8,22 @@
     top: 0;
     left: 0;
 }
+
+.popup{
+    position: relative;
+    z-index: 2;
+    display: flex;
+    width: 720px;
+    flex-direction: column;
+    align-items: center;
+    background: #1C1C21;
+    box-shadow: 0px 24px 32px rgba(0, 0, 0, 0.04), 0px 16px 24px rgba(0, 0, 0, 0.04), 0px 4px 8px rgba(0, 0, 0, 0.04), 0px 0px 1px rgba(0, 0, 0, 0.04);
+    border-radius: 40px;
+}
+
+.close{
+    position: absolute;
+    top: 60px;
+    right: 40px;
+    cursor: pointer;
+}
